% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_adjust.R
\name{aba_adjust}
\alias{aba_adjust}
\title{Create aba adjust object for p-value adjustment}
\usage{
aba_adjust(
  model_summary,
  method = c("bonferroni", "fdr", "hochberg", "holm", "hommel", "BH", "BY"),
  by = c("group", "outcome", "stat"),
  form = c("metric", "coef")
)
}
\arguments{
\item{method}{method}

\item{by}{by}

\item{form}{form}
}
\value{

}
\description{
Create aba adjust object for p-value adjustment
}
\examples{
x <- 1
}
