% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_ancova.R
\name{aba_ancova}
\alias{aba_ancova}
\title{Create an ancova stat to use for an aba model.}
\usage{
aba_ancova(std.beta = FALSE, complete.cases = TRUE)
}
\value{
list of the following functions:
\itemize{
\item \code{formula_fn}: create a formula
\item \code{fit_fn}: fit a model
\item \code{evaluate_fn}: evaluate a model
}
}
\description{
Create an ancova stat to use for an aba model.
}
\examples{
x <- 1
}
