% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_readwrite.R
\name{aba_board}
\alias{aba_board}
\title{Create an aba board}
\usage{
aba_board(path, ...)
}
\arguments{
\item{path}{filepath}

\item{...}{other}
}
\value{
abaBoard object
}
\description{
Create an aba board
}
\examples{
x <- 1
}
