% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_fit.R
\name{aba_compile}
\alias{aba_compile}
\title{Generic compile method}
\usage{
aba_compile(model)
}
\arguments{
\item{model}{aba-type model}
}
\value{
aba-type model
}
\description{
Generic compile method
}
\examples{
1 == 1
}
