% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_demo.R
\name{aba_demo}
\alias{aba_demo}
\title{Create a demographics table from a fitted aba object}
\usage{
aba_demo(
  model,
  strata = NULL,
  include_predictors = TRUE,
  include_covariates = TRUE,
  include_outcomes = TRUE,
  add_vars = NULL,
  data_filter = NULL,
  ...
)
}
\arguments{
\item{...}{additional parameters}

\item{object}{aba object}
}
\value{

}
\description{
Create a demographics table from a fitted aba object
}
\examples{
x <- 1
}
