% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{aba_emmeans}
\alias{aba_emmeans}
\title{Run emmeans on a model}
\usage{
aba_emmeans(model, treatment, stats_obj, ...)
}
\arguments{
\item{model}{model}

\item{treatment}{treatment}

\item{stats_obj}{stats obj}

\item{...}{extra parameters}
}
\value{

}
\description{
Run emmeans on a model
}
\examples{
x <- 1
}
