% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_fit.R
\name{aba_fit.abaModel}
\alias{aba_fit.abaModel}
\title{Fit an aba model.}
\usage{
\method{aba_fit}{abaModel}(object, ...)
}
\arguments{
\item{object}{abaModel. The aba model to be fitted.}

\item{...}{additional parameters.}
}
\value{
abaModel
}
\description{
This will trigger the fitting of all statistical models
(\code{stats}) on the different parameter combinations (\code{spec}).
}
\examples{
m <- aba_model()
}
