% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_fit.R
\name{aba_fit}
\alias{aba_fit}
\title{Generic fit method}
\usage{
aba_fit(object, ...)
}
\arguments{
\item{model}{aba-type model}
}
\value{
aba-type model
}
\description{
Generic fit method
}
\examples{
1 == 1
}
