% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_glm.R
\name{aba_glm}
\alias{aba_glm}
\title{Create a glm stat to use for an aba model.}
\usage{
aba_glm()
}
\value{
list of the following functions:
\itemize{
\item \code{formula_fn}: create a formula
\item \code{fit_fn}: fit a model
\item \code{evaluate_fn}: evaluate a model
}
}
\description{
Create a glm stat to use for an aba model.
}
\examples{
my_stat <- aba_glm()

my_formula <- my_stat$formula_fn(
  outcome='ConvertedToAlzheimers',
  predictors=c('PLASMA_PTAU181_bl','PLASMA_NFL_bl'),
  covariates=c('AGE_bl','GENDER','EDUCAT')
)

my_model <- my_stat$fit_fn(
  formula = my_formula,
  data = adni_sample
)
}
