% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_glm.R
\name{aba_glm_formula}
\alias{aba_glm_formula}
\title{Create a formula for glm from parameters}
\usage{
aba_glm_formula(outcome, predictors, covariates)
}
\arguments{
\item{outcome}{character. the outcome}

\item{predictors}{character vector. the predictors}

\item{covariates}{character vector. the covariates}
}
\value{
character. The formula to use when fitting the model
}
\description{
Create a formula for glm from parameters
}
