% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_lm.R
\name{aba_lm}
\alias{aba_lm}
\title{Create an lm stat to use for an aba model.}
\usage{
aba_lm(std.beta = FALSE, complete.cases = TRUE)
}
\value{
list of the following functions:
\itemize{
\item \code{formula_fn}: create a formula
\item \code{fit_fn}: fit a model
\item \code{evaluate_fn}: evaluate a model
}
}
\description{
Create an lm stat to use for an aba model.
}
\examples{
my_stat <- aba_lm()

my_formula <- my_stat$formula_fn(
  outcome='CDRSB_bl',
  predictors=c('PLASMA_PTAU181_bl','PLASMA_NFL_bl'),
  covariates=c('AGE','GENDER','EDUCATION')
)

my_model <- my_stat$fit_fn(
  formula = my_formula,
  data = aba::adnimerge \%>\% dplyr::filter(VISCODE == 'bl')
)
}
