% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_lme.R
\name{aba_lme}
\alias{aba_lme}
\title{Create a lme stat to use for an aba model.}
\usage{
aba_lme(id_var, time_var)
}
\value{
list of the following functions:
\itemize{
\item \code{formula_fn}: create a formula
\item \code{fit_fn}: fit a model
\item \code{evaluate_fn}: evaluate a model
}
}
\description{
Create a lme stat to use for an aba model.
}
\examples{
my_stat <- aba_lme()

my_formula <- my_stat$formula_fn(
  outcome='ConvertedToAlzheimers',
  predictors=c('PLASMA_PTAU181_bl','PLASMA_NFL_bl'),
  covariates=c('AGE_bl','GENDER','EDUCAT'),
  id
)

my_model <- my_stat$fit_fn(
  formula = my_formula,
  data = adni_sample
)
}
