% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_save.R
\name{aba_load}
\alias{aba_load}
\title{Load an ABA object from file}
\usage{
aba_load(file, ...)
}
\arguments{
\item{file}{ABA object}

\item{...}{filename}
}
\value{
NA
}
\description{
Load an ABA object from file
}
\examples{
x <- 1
}
