% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_longpower.R
\name{aba_longpower}
\alias{aba_longpower}
\title{Run power analysis on longitudinal data}
\usage{
aba_longpower(
  object,
  n = NULL,
  pct_change = NULL,
  power = NULL,
  t_length = NULL,
  t_freq = NULL,
  sig_level = 0.05,
  dropout = 0
)
}
\arguments{
\item{object}{object}

\item{n}{n}

\item{pct_change}{pct_change}

\item{power}{power}

\item{sig_level}{sig_level}

\item{t}{t}
}
\value{

}
\description{
Run power analysis on longitudinal data
}
\examples{
x <- 1
}
