% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_model.R
\name{aba_model}
\alias{aba_model}
\title{Create an aba model}
\usage{
aba_model(data = NULL, spec = aba_model_spec(), verbose = FALSE)
}
\arguments{
\item{data}{data.frame the data to use for the object}

\item{spec}{abaModelSpec the spec to use for the model. Can be created with
model_spec().}

\item{results}{list the fitted statistical models}
}
\value{
An abaModel object
}
\description{
An aba model is composed of the following:
\itemize{
\item data: a data.frame to be used to fit the statistical models
\item spec: the specification for the aba model composed of the following:
\itemize{
\item groups: which parts of the data to fit separate models on
\item outcomes: which variables to use as dependent variables
\item covariates: which variables to use as fixed independent variables in
every single model that is fit
\item predictors: which variables to use as independent variables, but never
together in the same model
}
\item fits: the fitted statistical models once \code{aba_fit()} is called
}
}
\examples{
m <- aba_model()
}
