% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_model_spec.R
\name{aba_model_spec}
\alias{aba_model_spec}
\title{Create a model spec}
\usage{
aba_model_spec(
  groups = "everyone",
  outcomes = NULL,
  covariates = NULL,
  predictors = NULL,
  stats = NULL
)
}
\arguments{
\item{groups}{vector. groups to use.}

\item{outcomes}{vector. outcomes to use.}

\item{covariates}{vector. covariates to use.}

\item{predictors}{vector or list. predictors to use}

\item{stats}{character or vector. stat to use.}
}
\value{
An abaModelSpec object
}
\description{
A model spec is composed of the following:
\itemize{
\item groups
\item outcomes
\item covariates
\item predictors
}
}
\examples{
spec <- aba_model_spec()
}
