% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_robust.R
\name{aba_plot.abaRobust}
\alias{aba_plot.abaRobust}
\title{Plot results of an aba robust object}
\usage{
\method{aba_plot}{abaRobust}(object, ...)
}
\arguments{
\item{object}{an aba robust object. The object whose results will be plotted.}
}
\value{
a ggplot
}
\description{
Plot results of an aba robust object
}
