% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_plot.R
\name{aba_plot}
\alias{aba_plot}
\title{Plot an aba object}
\usage{
aba_plot(object, ...)
}
\arguments{
\item{object}{aba object. The object to plot.}

\item{...}{additional parameters}
}
\value{
a ggplot with plotted results depending on the aba object.
}
\description{
This is a generic function for plotting an aba object. The resulting plot will
depend on the type of aba object. The supported objects are the following:
\itemize{
\item aba_emmeans
\item aba_longpower
\item aba_robust
}
}
