% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_plot.R
\name{aba_plot_coef.abaSummary}
\alias{aba_plot_coef.abaSummary}
\title{aba summary plot coef method}
\usage{
\method{aba_plot_coef}{abaSummary}(model_summary, ...)
}
\arguments{
\item{...}{additional parameters}

\item{object}{aba model summary}
}
\value{
ggplot
}
\description{
aba summary plot coef method
}
\examples{
x <- 1
}
