% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_plot.R
\name{aba_plot_coef}
\alias{aba_plot_coef}
\title{Generic plot coef method}
\usage{
aba_plot_coef(
  object,
  x = "predictor",
  group = "predictor_set",
  facet = c("outcome", "group"),
  coord_flip = FALSE,
  palette = "jama",
  labels = NULL,
  plotly = FALSE,
  ...
)
}
\arguments{
\item{object}{aba model summary}

\item{...}{additional parameters}
}
\value{
ggplot
}
\description{
Generic plot coef method
}
\examples{
x <- 1
}
