% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_robust.R
\name{aba_plot_metric.abaRobust}
\alias{aba_plot_metric.abaRobust}
\title{Plot AUC difference of aba robust object}
\usage{
\method{aba_plot_metric}{abaRobust}(object)
}
\arguments{
\item{object}{abaRobust. object to plot results from}
}
\value{
ggplot
}
\description{
Plot AUC difference of aba robust object
}
\examples{
x <- 1
}
