% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_plot.R
\name{aba_plot_metric}
\alias{aba_plot_metric}
\title{Plot performance metrics from an aba model}
\usage{
aba_plot_metric(model_summary, ...)
}
\arguments{
\item{...}{other plotting parameters}

\item{model}{abaModel. Fitted aba model to plot.}
}
\value{
ggplot or list of ggplots
}
\description{
Plot performance metrics from an aba model
}
\examples{
x <- 1
}
