% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_glm.R
\name{aba_plot_predictor_risk}
\alias{aba_plot_predictor_risk}
\title{Plot predictor values versus predicted risk from fitted aba model}
\usage{
aba_plot_predictor_risk(model, term_labels = NULL)
}
\arguments{
\item{model}{fitted abaModel. The model to plot}

\item{term_labels}{list. Names and values of predictors and their labels
to replace them with in the plot.}
}
\value{
a ggplot
}
\description{
This function plots real predictor values versus predicted risk (either
absolute or relative) from a fitted aba model with glm stats.
}
\examples{
data <- adnimerge \%>\% dplyr::filter(VISCODE == 'bl')
# fit glm model with binary outcome variables
model <- data \%>\% aba_model() \%>\%
  set_groups(everyone()) \%>\%
  set_outcomes(CSF_ABETA_STATUS_bl) \%>\%
  set_predictors(
    c(PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl)
  ) \%>\%
  set_stats(
    stat_glm(std.beta = FALSE)
  ) \%>\%
  fit()
g <- model \%>\% aba_plot_predictor_risk()
fig <- g$fig[[1]] # plot of interest
}
