% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_plot.R
\name{aba_plot_roc}
\alias{aba_plot_roc}
\title{Plot ROC curves from an aba model}
\usage{
aba_plot_roc(model, ...)
}
\arguments{
\item{model}{abaModel. Fitted aba model to plot.}

\item{...}{other plotting parameters}
}
\value{
gpglot or list of ggplots
}
\description{
Plot ROC curves from an aba model
}
\examples{
x <- 1
}
