% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_predict.R
\name{aba_predict}
\alias{aba_predict}
\title{Get individual predictions from a fitted aba model}
\usage{
aba_predict(model, newdata = NULL, include_data = TRUE)
}
\arguments{
\item{model}{a fitted aba model. The model to get predictions from.}

\item{newdata}{dataframe (optional). New data to get predictions from. If
this is null, predictions will be provided for the data originally used
to fit the aba model}

\item{include_data}{logical. Whether to also include the data with the
fitted/predicted values or just the fitted/predicted values for each fit.}
}
\value{
a dataframe with original data and with fitted/predicted values added
for each group-outcome-stat-predictor combination.
}
\description{
This function allows you to get the fitted/predicted values from data
used to fit an aba model or from new data. This is essentially an extension
of the \code{broom::augment()} function.
}
\examples{
data <- adnimerge \%>\% dplyr::filter(VISCODE == 'bl')

model <- data \%>\% aba_model() \%>\%
  set_groups(
    everyone(),
    DX_bl \%in\% c('MCI', 'AD')
  ) \%>\%
  set_outcomes(CDRSB_bl, MMSE_bl) \%>\%
  set_predictors(
    PLASMA_PTAU181_bl, PLASMA_NFL_bl,
    c(PLASMA_PTAU181_bl, PLASMA_NFL_bl)
  ) \%>\%
  set_covariates(AGE, GENDER, EDUCATION) \%>\%
  set_stats('lm')

model <- model \%>\% fit()
data_augmented <- model \%>\% aba_predict()
}
