% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_readwrite.R
\name{aba_read}
\alias{aba_read}
\title{Read an ABA object from file}
\usage{
aba_read(file, name = NULL, ...)
}
\arguments{
\item{file}{ABA object}

\item{...}{filename}
}
\value{
NA
}
\description{
Read an ABA object from file
}
\examples{
x <- 1
}
