% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_robust.R
\name{aba_robust}
\alias{aba_robust}
\title{Create an aba robust object}
\usage{
aba_robust(
  model,
  variation,
  variation_type = c("perc"),
  ntrials = 100,
  verbose = TRUE
)
}
\arguments{
\item{model}{abaModel the aba model to use for the object}

\item{variation}{list the test-retest (or whatever) variation estimates
for each predictor. This variability is assumed to standrd deviation of
of a normal distribution to be simulated.}

\item{variation_type}{character. what the variation values represent.
We only support percent change ("perc") right now.}

\item{ntrials}{integer number of noise simulations to run}
}
\value{
An abaRobust object
}
\description{
An aba robust object is composed of the following:
\itemize{
\item data: a data.frame to be used to fit the statistical models
\item spec: the specification for the aba model composed of the following:
\itemize{
\item groups: which parts of the data to fit separate models on
\item outcomes: which variables to use as dependent variables
\item covariates: which variables to use as fixed independent variables in
every single model that is fit
\item predictors: which variables to use as independent variables, but never
together in the same model
}
\item fits: the fitted statistical models once \code{fit()} is called
}
}
\examples{
x <- 1
}
