% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_save.R
\name{aba_save}
\alias{aba_save}
\title{Save an ABA object to file}
\usage{
aba_save(object, file, include_data = TRUE, include_fit = TRUE)
}
\arguments{
\item{object}{ABA object}

\item{file}{filename}

\item{...}{other params}
}
\value{
NA
}
\description{
Save an ABA object to file
}
\examples{
x <- 1
}
