% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_screen.R
\name{aba_screen}
\alias{aba_screen}
\title{aba (pre-)screening model}
\usage{
aba_screen(
  model,
  threshold,
  cost_multiplier,
  include_n,
  ntrials = 100,
  verbose = TRUE
)
}
\arguments{
\item{model}{aba model. should be fitted, only glm supported now.}

\item{threshold}{value or vector of values. between 0 and 1. Relative cutoff
from glm model to use as theoretical screening inclusion cutoff}

\item{ntrials}{integer. number of bootstrap trials.}

\item{verbose}{logical. whether to print updates.}
}
\value{

}
\description{
aba (pre-)screening model
}
\examples{
x <- 1
}
