% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_selection.R
\name{aba_selection}
\alias{aba_selection}
\title{Cfreate an aba selection object}
\usage{
aba_selection(
  model,
  method = c("forward", "backward"),
  criteria = c("aic", "pval"),
  threshold = NULL,
  verbose = FALSE
)
}
\arguments{
\item{model}{model}

\item{method}{method}

\item{criteria}{criteria}

\item{verbose}{verbose}
}
\value{

}
\description{
Cfreate an aba selection object
}
\examples{
x <- 1
}
