% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_summary.R
\name{aba_summary}
\alias{aba_summary}
\title{Create a summary of an aba model.}
\usage{
aba_summary(model, control = aba_control(), verbose = FALSE, ...)
}
\arguments{
\item{model}{abaModel. the model to create a summary from}

\item{...}{other params}
}
\value{
abaSummary object
}
\description{
Create a summary of an aba model.
}
\examples{
m <- aba_model()
}
