% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{aba_tidy}
\alias{aba_tidy}
\title{aba tidy generic}
\usage{
aba_tidy(model, predictors, covariates, ...)
}
\arguments{
\item{model}{model}

\item{predictors}{predictors}

\item{covariates}{covariates}

\item{...}{extra parameters}
}
\description{
aba tidy generic
}
