% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_grid.R
\name{aba_trajectory}
\alias{aba_trajectory}
\title{Calculate longitudinal trajectories from aba model}
\usage{
aba_trajectory(
  model,
  time_vals = NULL,
  predictor_mode = "mean",
  covariate_mode = "mean",
  level = 0
)
}
\arguments{
\item{level}{}
}
\description{
Note: This only works on longitudinal stats
}
