% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_trial.R
\name{aba_trial}
\alias{aba_trial}
\title{Create an aba trial}
\usage{
aba_trial(data = NULL, spec = aba_trial_spec(), results = list())
}
\arguments{
\item{data}{data.frame. the data to use for the object}

\item{spec}{trialSpec. the spec to use for the model. Can be created with
trial_spec()}

\item{results}{list the fitted statistical models}
}
\value{
An abaTrial object
}
\description{
An aba trial is composed of the following:
\itemize{
\item data: a data.frame to be used to fit the trial models
\item spec: the specification for the aba trial
\item fits: the fitted statistical models once \code{fit()} is called
}
}
\examples{
m <- aba_model()
}
