% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_trial_spec.R
\name{aba_trial_spec}
\alias{aba_trial_spec}
\title{Create a trial spec}
\usage{
aba_trial_spec(
  groups = NULL,
  endpoints = NULL,
  treatment = NULL,
  covariates = NULL,
  stats = NULL
)
}
\arguments{
\item{groups}{vector. group selection criteria to use.}

\item{endpoints}{vector. endpoints}

\item{treatment}{character. treatment}

\item{covariates}{vector. covariates}

\item{stats}{character or vector. stat to use.}
}
\value{
A trialSpec object
}
\description{
A trial spec is composed of the following:
\itemize{
\item groups
\item outcomes
\item times
\item stats
}
}
\examples{
spec <- aba_trial_spec()
}
