% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_readwrite.R
\name{aba_write.abaSummary}
\alias{aba_write.abaSummary}
\title{Write an aba summary object to file}
\usage{
\method{aba_write}{abaSummary}(object, filename, format = c("table", "raw", "object"), ...)
}
\description{
Write an aba summary object to file
}
