% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_readwrite.R
\name{aba_write}
\alias{aba_write}
\title{Write an ABA object to file}
\usage{
aba_write(object, file, name = NULL, include_data = TRUE, include_fit = TRUE)
}
\arguments{
\item{object}{ABA object}

\item{file}{filename}

\item{...}{other params}
}
\value{
NA
}
\description{
Write an ABA object to file
}
\examples{
x <- 1
}
