% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_readwrite.R
\name{aba_write.TableOne}
\alias{aba_write.TableOne}
\title{Write an aba demo table (TableOne) to file}
\usage{
\method{aba_write}{TableOne}(object, filename, ...)
}
\description{
Write an aba demo table (TableOne) to file
}
