% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_spec.R
\name{add_predictors}
\alias{add_predictors}
\title{Add predictors to an aba model}
\usage{
add_predictors(object, ...)
}
\arguments{
\item{object}{An aba model. The model for which you want to set predictors}

\item{...}{strings or variables or tidy-selection functions. Each
comma-separated value will be a new
predictor set. If you supply actual variables, then the data of the aba
model should already be set.}
}
\value{
An aba model with predictors added.
}
\description{
This function adds predictors to the already set predictors of an aba model.
}
\examples{
data <- adnimerge \%>\% dplyr::filter(VISCODE == 'bl')

model <- aba_model() \%>\%
  set_data(data) \%>\%
  set_predictors(
    PLASMA_ABETA_bl,
    PLASMA_PTAU181_bl,
    PLASMA_NFL_bl,
    c(PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl)
  )

model <- model \%>\%
  add_predictors(c(PLASMA_ABETA_bl, PLASMA_PTAU181_bl))

}
