% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adnimerge}
\alias{adnimerge}
\title{A sample of ADNI data in long format}
\format{
A data frame with 15,598 rows and 28 variables.
}
\source{
\url{https://adni.loni.usc.edu}
}
\usage{
adnimerge
}
\description{
A sample of ADNI data in long format. This data contains longitudinal
follow-up visits from CU and MCI patients, with baseline measures of
plasma biomarkers (abeta, p-tau, nfl), CSF biomarkers, MRI measures,
cognition, and demographics. Longitudinal outcomes include cognition
(MMSE, CDRSB) and conversion to AD dementia and conversion to
all-cause dementia. There is also binarized biomarker status for
many of the variables.
}
\keyword{datasets}
