% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_combos}
\alias{all_combos}
\title{Create all possible combinations of a set of variables}
\usage{
all_combos(...)
}
\arguments{
\item{...}{strings. Variable names to create combinations of}
}
\value{
list of vectors
}
\description{
Create all possible combinations of a set of variables
}
\examples{
x <- all_combos('a','b','c')
}
