% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_trial.R
\name{compile.abaTrial}
\alias{compile.abaTrial}
\title{Compile an abaTrial model}
\usage{
\method{compile}{abaTrial}(model)
}
\arguments{
\item{model}{abaTrial. model to compile}
}
\value{
abaTrial
}
\description{
Compile an abaTrial model
}
\examples{
m <- aba_trial()
}
