% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{compile}
\alias{compile}
\title{Compile the spec of an abaModel to ensure it is consistent with the
data that the abModel has. This also creates a tibble with all the
different combinations that will be tested ('fits')}
\usage{
compile(model)
}
\arguments{
\item{model}{abaModel. The model to compile.}
}
\value{
abaModel
}
\description{
Compile the spec of an abaModel to ensure it is consistent with the
data that the abModel has. This also creates a tibble with all the
different combinations that will be tested ('fits')
}
\examples{
m <- aba_model()
}
