% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_boot.R
\name{eval_boot}
\alias{eval_boot}
\title{Create a bootstrap evaluator}
\usage{
eval_boot(ntrials = 10)
}
\arguments{
\item{ntrials}{integer. number of trials to perform}
}
\value{
aba model
}
\description{
Create a bootstrap evaluator
}
\examples{
data <- adnimerge \%>\% dplyr::filter(VISCODE == 'bl')
model <- aba_model() \%>\%
  set_data(data) \%>\%
  set_groups(everyone()) \%>\%
  set_outcomes(ConvertedToAlzheimers, CSF_ABETA_STATUS_bl) \%>\%
  set_predictors(
    PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl,
    c(PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl)
  ) \%>\%
  set_stats('glm') \%>\%
  set_evals(eval_boot(ntrials = 5)) \%>\%
  fit()
}
