% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_cv.R
\name{eval_cv}
\alias{eval_cv}
\title{Create a cross validation evaluator}
\usage{
eval_cv(nfolds = 5, ntrials = 1)
}
\arguments{
\item{nfolds}{integer. number of cv folds}

\item{ntrials}{integer. number of cv trials to run}
}
\value{
aba model
}
\description{
Create a cross validation evaluator
}
\examples{
data <- adnimerge \%>\% dplyr::filter(VISCODE == 'bl')
model <- aba_model() \%>\%
  set_data(data) \%>\%
  set_groups(everyone()) \%>\%
  set_outcomes(ConvertedToAlzheimers, CSF_ABETA_STATUS_bl) \%>\%
  set_predictors(
    PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl,
    c(PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl)
  ) \%>\%
  set_stats('glm') \%>\%
  set_evals('cv') \%>\%
  fit()
}
