% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_robust.R
\name{fit.abaRobust}
\alias{fit.abaRobust}
\title{Fit an aba robust object}
\usage{
\method{fit}{abaRobust}(object, ...)
}
\arguments{
\item{object}{abaModel. The aba model whose robustness will be tested. This
aba model should already be fit itself prior to this.}

\item{...}{additional parameters.}
}
\value{
abaRobust
}
\description{
This will trigger the fitting of an aba robust object to determine the
effect of simulated noise/variation/test-retest on model coefficients
and preformance.
}
\examples{
x <- 1
}
