% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_screen.R
\name{fit.abaScreen}
\alias{fit.abaScreen}
\title{Fit an aba screen object}
\usage{
\method{fit}{abaScreen}(object, ...)
}
\arguments{
\item{object}{aba screen object to fit}

\item{...}{additional parameters}
}
\value{
abaScreen object
}
\description{
Fit an aba screen object
}
\examples{
x <- 1
}
