% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_selection.R
\name{fit.abaSelection}
\alias{fit.abaSelection}
\title{Fit an aba selection object}
\usage{
\method{fit}{abaSelection}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{additional parameters}
}
\value{

}
\description{
Fit an aba selection object
}
\examples{
x <- 1
}
