% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_trial.R
\name{fit.abaTrial}
\alias{fit.abaTrial}
\title{Fit an aba trial}
\usage{
\method{fit}{abaTrial}(object, ...)
}
\arguments{
\item{object}{abaTrial The aba trial to be fitted.}

\item{...}{additional parameters.}
}
\value{
abaTrial
}
\description{
This will trigger the fitting of all models
(\code{stats}) on the different parameter combinations (\code{spec}).
}
\examples{
m <- aba_trial()
}
