% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit an aba model. This will trigger the fitting of all statistical models
(\code{stats}) on the different parameter combinations (\code{spec}).}
\usage{
fit(model)
}
\arguments{
\item{model}{abaModel. The aba model to be fitted.}
}
\value{
abaModel
}
\description{
Fit an aba model. This will trigger the fitting of all statistical models
(\code{stats}) on the different parameter combinations (\code{spec}).
}
\examples{
m <- aba_model()
}
