% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_boot.R
\name{fit_boot}
\alias{fit_boot}
\title{Fit an aba model using bootstrap sampling}
\usage{
fit_boot(object, ntrials, verbose = FALSE)
}
\arguments{
\item{object}{aba model. The model to fit.}

\item{ntrials}{integer. Number of bootstrap trials to run.}

\item{verbose}{logical. Whether to print updates using progress bar.}
}
\value{
a fitted aba model
}
\description{
This function fits an aba model first on the full dataset and then
using bootrstrap sampling.
}
\examples{
data <- adnimerge \%>\% dplyr::filter(VISCODE == 'bl')

model_spec <- aba_model() \%>\%
  set_data(data) \%>\%
  set_groups(everyone()) \%>\%
  set_outcomes(ConvertedToAlzheimers, CSF_ABETA_STATUS_bl) \%>\%
  set_predictors(
    PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl,
    c(PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl)
  ) \%>\%
  set_stats('glm')

model <- model_spec \%>\% fit_boot(ntrials = 5)
}
