% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cv.R
\name{fit_cv}
\alias{fit_cv}
\title{Fit an aba model using cross validation sampling}
\usage{
fit_cv(object, nsplits = 5, ntrials = 1, verbose = FALSE)
}
\arguments{
\item{object}{aba model. The model to fit.}

\item{nsplits}{integer. Number of equal-sized cross validation splits to use.}

\item{ntrials}{integer. Number of trials to run.}

\item{verbose}{logical. Whether to print updates using progress bar.}
}
\value{
a fitted aba model
}
\description{
This function fits an aba model on cross validation training data splits and
stores the test data for evaluation.
}
\examples{
data <- adnimerge \%>\% dplyr::filter(VISCODE == 'bl')

model <- aba_model() \%>\%
  set_data(data) \%>\%
  set_groups(everyone()) \%>\%
  set_outcomes(ConvertedToAlzheimers, CSF_ABETA_STATUS_bl) \%>\%
  set_predictors(
    PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl,
    c(PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl)
  ) \%>\%
  set_stats('glm')

model <- model \%>\% fit_cv(nsplits = 3, ntrials = 5)
}
