% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_traintest.R
\name{fit_traintest}
\alias{fit_traintest}
\title{Fit an aba model using train-test sampling}
\usage{
fit_traintest(object, split = 0.8, ntrials = 1, verbose = FALSE)
}
\arguments{
\item{object}{aba model. The model to fit.}

\item{split}{double between 0 and 1. Percent of data to sample for train set.}

\item{ntrials}{integer. Number of trials to run.}

\item{verbose}{logical. Whether to print updates using progress bar.}
}
\value{
a fitted aba model
}
\description{
This function fits an aba model on training data and stores the test data
for evaluation.
}
\examples{
data <- adnimerge \%>\% dplyr::filter(VISCODE == 'bl')

model <- aba_model() \%>\%
  set_data(data) \%>\%
  set_groups(everyone()) \%>\%
  set_outcomes(ConvertedToAlzheimers, CSF_ABETA_STATUS_bl) \%>\%
  set_predictors(
    PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl,
    c(PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl)
  ) \%>\%
  set_stats('glm')

model <- model \%>\% fit_traintest(split = 0.8, ntrials = 5)
}
