% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_spec.R
\name{get_predictors}
\alias{get_predictors}
\title{Get all predictors of an aba model.}
\usage{
get_predictors(object)
}
\arguments{
\item{object}{an ab model. The model to get predictors from.}
}
\value{
A vector of strings containing all unique predictor variables.
}
\description{
This function can be used to get all unique predictor variables from an
aba model. This can be thought of as "flattening" all of the predictor
sets of an aba model.
}
\examples{

data <- adnimerge \%>\% dplyr::filter(VISCODE == 'bl')

model <- aba_model() \%>\%
  set_data(data) \%>\%
  set_predictors(
    PLASMA_ABETA_bl,
    PLASMA_PTAU181_bl,
    PLASMA_NFL_bl,
    c(PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl)
  )

predictors <- model \%>\% get_predictors()

}
