% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{run_emmeans}
\alias{run_emmeans}
\title{Run emmeans on a model}
\usage{
run_emmeans(fit, extra_params)
}
\arguments{
\item{model}{model}

\item{treatment}{treatment}

\item{stats_obj}{stats obj}

\item{...}{extra parameters}
}
\value{

}
\description{
Run emmeans on a model
}
\examples{
x <- 1
}
