% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_selection.R
\name{run_selection}
\alias{run_selection}
\title{Fit an aba selection object}
\usage{
run_selection(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{additional parameters}
}
\value{

}
\description{
Fit an aba selection object
}
\examples{
x <- 1
}
