% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_data.R
\name{set_data}
\alias{set_data}
\title{Set data element of an aba model}
\usage{
set_data(.model, ...)
}
\arguments{
\item{.model}{abaModel. model to which data will be added}

\item{...}{data.frames. data to add}
}
\value{
An abaModel object
}
\description{
Set data element of an aba model
}
\examples{
m <- aba_model() \%>\% set_data(data.frame(x=c(1,2,3)))
}
