% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_trial_spec.R
\name{set_endpoints}
\alias{set_endpoints}
\title{Set the endpoints of an aba_trial spec}
\usage{
set_endpoints(.model, ...)
}
\arguments{
\item{.model}{abaTrial}

\item{...}{endpoints}
}
\value{
An abaTrial object
}
\description{
Set the endpoints of an aba_trial spec
}
\examples{
m <- aba_model() \%>\% set_outcomes()
}
