% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_utils.R
\name{set_evals}
\alias{set_evals}
\title{Set the evals of an aba model}
\usage{
set_evals(.model, ..., .labels = NULL)
}
\arguments{
\item{.model}{aba model. The model to set the evals for.}

\item{...}{comma-separated strings or abaEval objects. The evals you wish
to set for the given model}

\item{.labels}{vector of strings (optional). The .labels for each eval.}
}
\value{
aba model with evals set.
}
\description{
Evals are the ways in which you fit your stats on the data. The standard
method is to simply fit the model on the entire dataset one time. However,
there are additional methods such as bootstrap sampling, train-test splits,
and cross validation. An aba model can have multiple evals, which can
be useful if you want to test the difference between cross validation with
five splits or three, for example.
}
\examples{
model <- aba_model() \%>\%
  set_evals(
    'standard'
    #'boot',
    #'traintest',
    #'cv'
  )
model <- aba_model() \%>\%
  set_evals(
    eval_standard()
    #eval_boot(ntrials = 100),
    #eval_traintest(split = 0.5, ntrials = 10),
    #eval_cv(nfolds = 3, ntrials = 10)
  )
model <- aba_model() \%>\%
  set_evals(
    eval_cv(nfolds = 3)
  )
}
