% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_spec.R
\name{set_inclusion}
\alias{set_inclusion}
\title{Set the inclusion of a trial spec}
\usage{
set_inclusion(.model, ...)
}
\arguments{
\item{.model}{abaTrial}

\item{...}{statements. inclusion criteria}
}
\value{
An abaTrial object
}
\description{
Set the inclusion of a trial spec
}
\examples{
m <- aba_model() \%>\% set_inclusion()
}
