% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_model_spec.R
\name{set_outcomes}
\alias{set_outcomes}
\title{Set the outcomes of an model spec}
\usage{
set_outcomes(.model, ..., labels = NULL)
}
\arguments{
\item{.model}{abaModel}

\item{...}{outcomes}

\item{labels}{vector optional labels for spec parameter - printing/plotting.}
}
\value{
An abaModel object
}
\description{
Set the outcomes of an model spec
}
\examples{
m <- aba_model() \%>\% set_outcomes()
}
