% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_model_spec.R
\name{set_predictors}
\alias{set_predictors}
\title{Set the predictors of an model spec}
\usage{
set_predictors(.model, ..., labels = NULL)
}
\arguments{
\item{.model}{abaModel}

\item{...}{predictors}

\item{labels}{vector optional labels for spec parameter - printing/plotting.}
}
\value{
An abaModel object
}
\description{
Set the predictors of an model spec
}
\examples{
m <- aba_model() \%>\% set_predictors()
}
