% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_spec.R
\name{set_stats}
\alias{set_stats}
\title{Set statistical model of an aba model}
\usage{
set_stats(model, stat)
}
\arguments{
\item{model}{abaModel. aba model to alter.}

\item{stat}{string. which statistical model to use.}
}
\value{
An abaModel object
}
\description{
Set statistical model of an aba model
}
\examples{
m <- aba_model() \%>\% set_stats('glm')
}
