% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_spec.R
\name{set_timepoints}
\alias{set_timepoints}
\title{Set the timepoints of a trial spec}
\usage{
set_timepoints(.model, time_var, ...)
}
\arguments{
\item{.model}{abaTrial}

\item{time_var}{variable or string. data variable with time values}

\item{...}{numeric or character. how long the trial will be.}
}
\value{
An abaTrial object
}
\description{
Set the timepoints of a trial spec
}
\examples{
m <- aba_model() \%>\% set_timepoints('VISIT', 1.5, 2)
}
