% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_spec.R
\name{set_times}
\alias{set_times}
\title{Set the times of a trial spec}
\usage{
set_times(.model, ...)
}
\arguments{
\item{.model}{abaTrial}

\item{...}{numeric or character. how long the trial will be.}
}
\value{
An abaTrial object
}
\description{
Set the times of a trial spec
}
\examples{
m <- adni_sample \%>\% aba_trial() \%>\%
  set_times(VISIT == 1.5, VISIT==2)
}
