% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_model_spec.R
\name{set_treatment}
\alias{set_treatment}
\title{Set the endpoints of an aba_trial spec}
\usage{
set_treatment(.model, ...)
}
\arguments{
\item{.model}{abaTrial}

\item{...}{endpoints}
}
\value{
An abaModel object
}
\description{
Set the endpoints of an aba_trial spec
}
\examples{
x <- 1
}
