% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cox.R
\name{stat_cox}
\alias{stat_cox}
\title{Create a glm stat object.}
\usage{
stat_cox(time, std.beta = FALSE, complete.cases = TRUE)
}
\arguments{
\item{time}{string. The "time under risk" variable determining how long e.g.
the individual has been in the study or when the individual got the disease.}

\item{std.beta}{logical. Whether to standardize model predictors and
covariates prior to analysis.}

\item{complete.cases}{logical. Whether to only include the subset of data
with no missing data for any of the outcomes, predictors, or covariates.
Note that complete cases are considering within each group - outcome
combination but across all predictor sets.}
}
\value{
An abaStat object with \code{glm} stat type.
}
\description{
This function creates a glm stat object which can be passed as input
to the \code{set_stats()} function when building an aba model. This stat performs
a traditional logistic regression analysis using the \code{glm} function with
a binary outcome. Coefficients will be presented as odds ratios. Default
metrics include AUC.
}
\examples{
data <- adnimerge \%>\% dplyr::filter(VISCODE == 'bl')

# fit model
model <- data \%>\% aba_model() \%>\%
  set_groups(everyone()) \%>\%
  set_outcomes(ConvertedToAlzheimers, ConvertedToDementia) \%>\%
  set_predictors(
    PLASMA_PTAU181_bl, PLASMA_NFL_bl,
    c(PLASMA_PTAU181_bl, PLASMA_NFL_bl)
  ) \%>\%
  set_stats(
    stat_cox(time = 'TimeUnderRiskDementia')
  ) \%>\%
  fit()

## summarise model
model_summary <- model \%>\% summary()
}
