% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_glm.R
\name{stat_glm}
\alias{stat_glm}
\title{Create a glm stat to use for an aba model.}
\usage{
stat_glm(
  std.beta = FALSE,
  complete.cases = TRUE,
  include.basic = TRUE,
  extra.metrics = NULL
)
}
\value{
list of the following functions:
\itemize{
\item \code{formula_fn}: create a formula
\item \code{fit_fn}: fit a model
\item \code{evaluate_fn}: evaluate a model
}
}
\description{
Create a glm stat to use for an aba model.
}
\examples{
my_stat <- stat_glm()

my_formula <- my_stat$formula_fn(
  outcome='ConvertedToAlzheimers',
  predictors=c('PLASMA_PTAU181_bl','PLASMA_NFL_bl'),
  covariates=c('AGE','GENDER','EDUCATION')
)

my_model <- my_stat$fit_fn(
  formula = my_formula,
  data = aba::adnimerge \%>\% dplyr::filter(VISCODE == 'bl')
)
}
