% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_mmrm.R
\name{stat_mmrm}
\alias{stat_mmrm}
\title{Create a mmrm stat to use for an aba model.}
\usage{
stat_mmrm(
  id,
  time,
  treatment = NULL,
  add_baseline = TRUE,
  std.beta = FALSE,
  complete.cases = TRUE
)
}
\arguments{
\item{id}{string. id variable}

\item{time}{string. time variable
(should be discrete visits common to all participants)}
}
\value{
list of the following functions:
\itemize{
\item \code{formula_fn}: create a formula
\item \code{fit_fn}: fit a model
\item \code{evaluate_fn}: evaluate a model
}
}
\description{
Create a mmrm stat to use for an aba model.
}
\examples{
my_stat <- stat_mmrm(id='SUBJECT_ID',
                    time='Years_bl')

#my_formula <- my_stat$formula_fn(
#  outcome='ConvertedToAlzheimers',
#  predictors=c('PLASMA_PTAU181_bl','PLASMA_NFL_bl'),
#  covariates=c('AGE_bl','GENDER','EDUCAT'),
#  id
#)

#my_model <- my_stat$fit_fn(
#  formula = my_formula,
#  data = adni_sample
#)
}
