% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_power.R
\name{stat_power}
\alias{stat_power}
\title{Create a power stat}
\usage{
stat_power(
  n = NULL,
  delta = NULL,
  power = NULL,
  alpha = 0.05,
  method = c("t.test", "lme")
)
}
\arguments{
\item{n}{integer. sample size}

\item{delta}{float. treatment effect between 0 - 1}

\item{power}{float. power between 0 - 1}

\item{alpha}{float. alpha value}

\item{method}{string. method used to calculate power}
}
\value{
abaStat object
}
\description{
Create a power stat
}
\examples{
p <- stat_power(alpha=0.05, power=0.8, delta=0.3, method='t.test')
}
