% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_roc.R
\name{stat_roc}
\alias{stat_roc}
\title{Create a roc stat to use for an aba model.}
\usage{
stat_roc(
  direction = ">",
  method = "Youden",
  tag.healthy = 0,
  std.beta = FALSE,
  complete.cases = TRUE,
  extra.metrics = NULL
)
}
\value{
list of the following functions:
\itemize{
\item \code{formula_fn}: create a formula
\item \code{fit_fn}: fit a model
\item \code{evaluate_fn}: evaluate a model
}
}
\description{
Create a roc stat to use for an aba model.
}
\examples{
my_stat <- stat_roc()

my_formula <- my_stat$formula_fn(
  outcome = 'ConvertedToAlzheimers',
  predictors = 'PLASMA_PTAU181_bl'
)

my_model <- my_stat$fit_fn(
  formula = my_formula,
  data = aba::adnimerge \%>\% dplyr::filter(VISCODE == 'bl'),
  extra_params = my_stat$extra_params
)
}
