% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_plot.R
\name{theme_aba}
\alias{theme_aba}
\title{Custom aba ggplot2 theme}
\usage{
theme_aba(
  base_size = 16,
  legend.position = "top",
  coord_flip = FALSE,
  legend_title = FALSE,
  facet_labels = TRUE,
  axis_title = FALSE,
  family = c("Tahoma", "Helvetica", "Verdana")
)
}
\arguments{
\item{base_size}{integer. base font size}

\item{legend.position}{string. where to place legend ('none' = no legend)}

\item{coord_flip}{logical. whether to flip x and y axes}

\item{legend_title}{logical. whether to include legend title.}

\item{facet_labels}{logical. Whether to include labels of facets.}

\item{axis_title}{logical. Whether to include axis title for non-primary axis}

\item{family}{string. which font family to use. Should be one of
Verdana, Tahoma, Helvetica}
}
\value{
ggplot2 theme which can be added to a ggplot object
}
\description{
Custom aba ggplot2 theme
}
\examples{
library(ggplot2)

data <- data.frame(x=1:10, y=1:10)
fig <- ggplot(data, aes(x=x,y=y)) + geom_point() + theme_aba()
}
