% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_spec.R
\name{trial_spec}
\alias{trial_spec}
\title{Create a trial spec}
\usage{
trial_spec(groups = NULL, outcomes = NULL, times = NULL, stats = NULL)
}
\arguments{
\item{groups}{vector. group selection criteria to use.}

\item{outcomes}{vector. outcomes to use.}

\item{times}{vector. times to use.}

\item{stats}{character or vector. stat to use.}
}
\value{
A trialSpec object
}
\description{
A trial spec is composed of the following:
\itemize{
\item groups
\item outcomes
\item times
\item stats
}
}
\examples{
spec <- trial_spec()
}
