% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba-package.R
\docType{package}
\name{aba}
\alias{aba}
\alias{aba-package}
\title{Automated biomarker analysis in R}
\description{
A tool to fit clinical prediction models and plan clinical trials using biomarker data across multiple analysis factors (groups, outcomes, predictors).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ncullen93/abaR/}
  \item \url{https://ncullen93.github.io/abaR/}
}

}
\author{
\strong{Maintainer}: Nicholas Cullen \email{nickcullen31@gmail.com}

}
\keyword{internal}
