% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_control.R
\name{aba_control}
\alias{aba_control}
\title{Create an aba control which determines how an aba summary will be printed.}
\usage{
aba_control(include_covariates = TRUE)
}
\arguments{
\item{covars}{boolean. Whether to include covariates summary.}
}
\value{

}
\description{
Create an aba control which determines how an aba summary will be printed.
}
\examples{
ctrl <- aba_control()
}
