% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_emmeans.R
\name{aba_emmeans}
\alias{aba_emmeans}
\title{Calculate estimated marginal / least-square means and treatment effects
for mmrm, lme, etc}
\usage{
aba_emmeans(model, ...)
}
\arguments{
\item{model}{abaModel}

\item{...}{additional parameters}

\item{pairs}{logical}
}
\value{

}
\description{
Calculate estimated marginal / least-square means and treatment effects
for mmrm, lme, etc
}
\examples{
x <- 1
}
