% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_evaluate.R
\name{aba_evaluate}
\alias{aba_evaluate}
\title{Evaluate a fitted aba model on new data}
\usage{
aba_evaluate(model, data)
}
\arguments{
\item{model}{a fitted aba model. The model to test on new data}

\item{data}{dataframe. The new data on which the fitted model will be tested.}
}
\value{
an object with class 'abaEvaluate' which contains model performance
on the data
}
\description{
This function lets you evaluate the performance of a fitted aba model on
new data.
}
\examples{
# create separate training and testing datasets
train_id <- sample(nrow(data), 0.8*nrow(data), replace = FALSE)
data_train <- adnimerge \%>\%
  dplyr::filter(VISCODE == 'bl', row_number() \%in\% train_id)
data_test <- data \%>\%
  dplyr::filter(VISCODE == 'bl', !row_number() \%in\% train_id)

# fit model on train data
model <- model \%>\%
  set_data(data_train) \%>\%
  fit()

# evaluate fitted model on test data
res <- model \%>\% aba_evaluate(data = data_test)
}
