% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_plot.R
\name{aba_plot_metric.abaSummary}
\alias{aba_plot_metric.abaSummary}
\title{Plot performance metrics from an aba model}
\usage{
\method{aba_plot_metric}{abaSummary}(
  object,
  metric = NULL,
  x = "predictor_set",
  group = "outcome",
  facet = "group",
  coord_flip = FALSE,
  palette = "jama",
  labels = NULL,
  plotly = FALSE,
  ...
)
}
\arguments{
\item{...}{other plotting parameters}

\item{model}{abaModel. Fitted aba model to plot.}
}
\value{
ggplot or list of ggplots
}
\description{
Plot performance metrics from an aba model
}
\examples{
x <- 1
}
