% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_plot.R
\name{aba_plot_metric}
\alias{aba_plot_metric}
\title{Generic plot metric method}
\usage{
aba_plot_metric(
  object,
  metric = NULL,
  x = "predictor_set",
  group = "outcome",
  facet = "group",
  coord_flip = FALSE,
  palette = "jama",
  labels = NULL,
  plotly = FALSE,
  ...
)
}
\arguments{
\item{object}{aba model summary}

\item{...}{additional parameters}
}
\value{
ggplot
}
\description{
Generic plot metric method
}
\examples{
x <- 1
}
