% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_glm.R
\name{aba_plot_risk_density}
\alias{aba_plot_risk_density}
\title{Plot risk density split by binary outcome class}
\usage{
aba_plot_risk_density(
  model,
  risk_type = c("absolute", "relative"),
  include_basic = TRUE
)
}
\arguments{
\item{model}{abaModel. A fitted aba model with \code{stat_glm} type.}
}
\value{
a ggplot with risk density curves split by binary outcome value.
}
\description{
This function plots risk density curves across group - outcome - stat
combinations and split by binary outcome. It currently supports \code{stat_glm}.
}
\examples{
data <- adnimerge \%>\% dplyr::filter(VISCODE == 'bl')

# fit glm model with binary outcome variables
model <- data \%>\% aba_model() \%>\%
  set_groups(everyone()) \%>\%
  set_outcomes(ConvertedToAlzheimers, CSF_ABETA_STATUS_bl) \%>\%
  set_predictors(
    PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl,
    c(PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl)
  ) \%>\%
  set_stats(
    stat_glm(std.beta = TRUE)
  ) \%>\%
  fit()

fig <- model \%>\% aba_plot_riskdensity()
}
