% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_glm.R
\name{aba_plot_roc}
\alias{aba_plot_roc}
\title{Plot ROC curves from an aba model}
\usage{
aba_plot_roc(model_summary, ...)
}
\arguments{
\item{...}{other plotting parameters}

\item{model}{abaModel. Fitted aba model to plot.}
}
\value{
gpglot or list of ggplots
}
\description{
Plot ROC curves from an aba model
}
\examples{
x <- 1
}
