% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_grid.R
\name{aba_plot_trajectory}
\alias{aba_plot_trajectory}
\title{Plot trajectories of fitted longitudinal stats}
\usage{
aba_plot_trajectory(model, time_vals = NULL, level = 0)
}
\arguments{
\item{level}{}
}
\description{
level=0: group-level estimates
level=1: individual-level estimates
level=2: individual-level observations
}
