% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_spec.R
\name{aba_spec}
\alias{aba_spec}
\title{Create an aba spec}
\usage{
aba_spec()
}
\value{
An abaSpec object
}
\description{
An aba spec is composed of the following:
\itemize{
\item groups
\item outcomes
\item covariates
\item predictors
}
}
\examples{
spec <- aba_spec()
}
