% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_readwrite.R
\name{aba_write}
\alias{aba_write}
\title{Write an ABA object to file}
\usage{
aba_write(object, filename, ...)
}
\arguments{
\item{object}{ABA object}

\item{name}{character. name}

\item{include_data}{logical. include data}

\item{include_fit}{logical. include fit}

\item{file}{character. filename}
}
\value{
NA
}
\description{
Write an ABA object to file
}
\examples{
x <- 1
}
