% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_model_spec.R
\name{add_predictors}
\alias{add_predictors}
\title{Add  predictors to an model spec}
\usage{
add_predictors(.model, ...)
}
\arguments{
\item{.model}{abaModel}

\item{...}{predictors}
}
\value{
An abaModel object
}
\description{
Add  predictors to an model spec
}
\examples{
m <- aba_model() \%>\% set_predictors()
}
