% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_traintest.R
\name{eval_traintest}
\alias{eval_traintest}
\title{Create a train-test evaluator}
\usage{
eval_traintest(split = 0.8, ntrials = 1)
}
\arguments{
\item{split}{double between 0 and 1. percent of data to use as train set}

\item{ntrials}{integer. number of train-test trials to run}
}
\value{
aba model
}
\description{
Create a train-test evaluator
}
\examples{
data <- adnimerge \%>\% dplyr::filter(VISCODE == 'bl')
model <- aba_model() \%>\%
  set_data(data) \%>\%
  set_groups(everyone()) \%>\%
  set_outcomes(ConvertedToAlzheimers, CSF_ABETA_STATUS_bl) \%>\%
  set_predictors(
    PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl,
    c(PLASMA_ABETA_bl, PLASMA_PTAU181_bl, PLASMA_NFL_bl)
  ) \%>\%
  set_stats('glm') \%>\%
  set_evals('traintest') \%>\%
  fit()
}
