% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_model_spec.R
\name{get_predictors}
\alias{get_predictors}
\title{Get all unique predictors}
\usage{
get_predictors(model)
}
\arguments{
\item{model}{abaModel. model to get predictors from}
}
\value{
character vector. all unique predictors
}
\description{
Get all unique predictors
}
\examples{
x <- 1
}
