% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_spec.R
\name{set_covariates}
\alias{set_covariates}
\title{Set the covariates of an aba spec}
\usage{
set_covariates(.model, ...)
}
\arguments{
\item{.model}{abaModel}

\item{...}{covariates}
}
\value{
An abaModel object
}
\description{
Set the covariates of an aba spec
}
\examples{
m <- aba_model() \%>\% set_covariates()
}
