% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_spec.R
\name{set_groups}
\alias{set_groups}
\title{Set the groups of an aba spec}
\usage{
set_groups(.model, ...)
}
\arguments{
\item{.model}{abaModel}

\item{...}{groups}
}
\value{
An abaModel with groups altered
}
\description{
Set the groups of an aba spec
}
\examples{
m <- aba_model() \%>\% set_groups()
}
