% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{set_outcomes}
\alias{set_outcomes}
\title{Set the outcomes of an model spec}
\usage{
set_outcomes(.model, ...)
}
\arguments{
\item{.model}{abaModel}

\item{...}{outcomes}
}
\value{
An abaModel object
}
\description{
Set the outcomes of an model spec
}
\examples{
m <- aba_model() \%>\% set_outcomes()
}
