% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_spec.R
\name{set_predictors}
\alias{set_predictors}
\title{Set the predictors of an aba spec}
\usage{
set_predictors(.model, ...)
}
\arguments{
\item{.model}{abaModel}

\item{...}{predictors}
}
\value{
An abaModel object
}
\description{
Set the predictors of an aba spec
}
\examples{
m <- aba_model() \%>\% set_predictors()
}
