% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aba_spec.R
\name{set_stats}
\alias{set_stats}
\title{Set statistical model of an aba model}
\usage{
set_stats(.model, ...)
}
\arguments{
\item{...}{vector. Which statistical models to use.}

\item{model}{abaModel. aba model to alter.}
}
\value{
An abaModel object
}
\description{
Set statistical model of an aba model
}
\examples{
# create default stat object by string
m <- aba_model() \%>\% set_stats('glm')
# create a stat object with - useful w/ extra params
m <- aba_model() \%>\% set_stats(aba_glm())
}
