% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_spec.R
\name{trial_spec}
\alias{trial_spec}
\title{Create a trial spec}
\usage{
trial_spec(inclusion = NULL, outcomes = NULL, timepoints = NULL, stats = NULL)
}
\arguments{
\item{inclusion}{vector. inclusion criteria to use.}

\item{outcomes}{vector. outcomes to use.}

\item{timepoints}{vector. timepoints to use.}

\item{stats}{character or vector. stat to use.}
}
\value{
A trialSpec object
}
\description{
A trial spec is composed of the following:
\itemize{
\item inclusion
\item outcomes
\item timepoints
\item stats
}
}
\examples{
spec <- aba_spec()
}
